# docs/API-CONTRACT.openapi.yaml
openapi: 3.0.3
info:
  title: Nursery Management System API
  version: 0.2.0
  description: |
    Canonical OpenAPI contract for the Nursery Management System API.
    Aligned with current Gin route registration in `api/internal/app/bootstrap/bootstrap.go` and module handlers.
    Includes operational/debug routes and deferred proposed contracts.
  contact:
    name: Nursery API
servers:
  - url: /
    description: Same-origin (dev/proxy)
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: "Authorization: Bearer <access_token>"
    refreshTokenCookie:
      type: apiKey
      in: cookie
      name: refresh_token
      description: HttpOnly cookie set after login/refresh
    csrfCookie:
      type: apiKey
      in: cookie
      name: csrf_token
      description: Readable cookie containing CSRF token value
    csrfHeader:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: Must match csrf_token cookie value
    stripeSignature:
      type: apiKey
      in: header
      name: Stripe-Signature
      description: Stripe webhook signature header
...
  schemas:
    ErrorResponse:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: string
          description: Machine-readable error code
        message:
          type: string
          description: Human-readable error message
        request_id:
          type: string
          format: uuid
        details:
          description: Optional field-level details
          nullable: true
          type: object
          additionalProperties: true
      example:
        code: validation_error
        message: Invalid request payload.
        request_id: "00000000-0000-0000-0000-000000000001"
        details:
          field: email
...
    Child:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        date_of_birth:
          type: string
          format: date
        start_date:
          type: string
          format: date
        end_date:
          type: string
          format: date
          nullable: true
        core_hourly_rate_minor:
          type: integer
        notes:
          type: string
          nullable: true
        is_active:
          type: boolean
        left_at:
          type: string
          format: date-time
          nullable: true
        left_reason_code:
          type: string
          nullable: true
        left_reason_note:
          type: string
          nullable: true
        enrollment_complete:
          type: boolean
        missing_requirements:
          type: array
          items:
            type: string
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
...
    AttendanceChild:
      type: object
      properties:
        id:
          type: string
          format: uuid
        full_name:
          type: string
        enrollment_complete:
          type: boolean
        attendance_state:
          type: string
          enum: [not_checked_in, checked_in, absent]
        open_session_id:
          type: string
          format: uuid
          nullable: true
        checked_in_at:
          type: string
          format: date-time
          nullable: true
        has_incomplete_session:
          type: boolean
        absence_marker_id:
          type: string
          format: uuid
          nullable: true
        absence_marked_at:
          type: string
          format: date-time
          nullable: true
...
    ReasonRequest:
      type: object
      required: [reason_code]
      properties:
        reason_code:
          type: string
          enum:
            [
              duplicate_record,
              entered_in_error,
              left_nursery,
              safeguarding_direction,
              contact_update,
              access_revoked,
              other,
            ]
        reason_note:
          type: string
          description: Required when reason_code is other
...
  responses:
    BadRequest:
      description: Validation error
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: validation_error
            message: Invalid request payload.
            request_id: "00000000-0000-0000-0000-000000000001"
            details:
              field: email
    Unauthorized:
      description: Missing or invalid credentials
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: unauthorized
            message: Invalid credentials or session.
            request_id: "00000000-0000-0000-0000-000000000001"
    ForbiddenRole:
      description: Authenticated but wrong role
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: forbidden_role
            message: Access denied.
            request_id: "00000000-0000-0000-0000-000000000001"
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: not_found
            message: Resource not found.
            request_id: "00000000-0000-0000-0000-000000000001"
...
  parameters:
    child_id:
      name: child_id
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
paths:
...
  /api/v1/children:
    get:
      operationId: listChildren
      summary: List children
      tags: [Children]
      x-roles: [manager]
      security:
        - bearerAuth: []
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [active, inactive, all]
            default: active
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
            default: 50
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
            default: 0
      responses:
        "200":
          description: Children list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/Child"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"

    post:
      operationId: createChild
      summary: Create child
      tags: [Children]
      x-roles: [manager]
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [full_name, date_of_birth, start_date, core_hourly_rate_minor]
              properties:
                full_name:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                core_hourly_rate_minor:
                  type: integer
                notes:
                  type: string
      responses:
        "201":
          description: Child created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
  /api/v1/children/attendance:
    get:
      operationId: listAttendanceChildren
      summary: List children with current attendance state
      tags: [Children, Attendance]
      x-roles: [manager, practitioner]
      security:
        - bearerAuth: []
      responses:
        "200":
          description: Children with attendance state for current day
          content:
            application/json:
              schema:
                type: object
                properties:
                  items:
                    type: array
                    items:
                      $ref: "#/components/schemas/AttendanceChild"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
  /api/v1/children/{child_id}:
    get:
      operationId: getChild
      summary: Get child by ID
      tags: [Children]
      x-roles: [manager]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/child_id"
      responses:
        "200":
          description: Child detail
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"

    patch:
      operationId: updateChild
      summary: Partial update child
      tags: [Children]
      x-roles: [manager]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/child_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                full_name:
                  type: string
                date_of_birth:
                  type: string
                  format: date
                start_date:
                  type: string
                  format: date
                end_date:
                  type: string
                  format: date
                core_hourly_rate_minor:
                  type: integer
                notes:
                  type: string
      responses:
        "200":
          description: Updated child
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
  /api/v1/children/{child_id}/actions/mark-inactive:
    post:
      operationId: markChildInactive
      summary: Mark child as inactive
      tags: [Children]
      x-roles: [manager]
      security:
        - bearerAuth: []
      parameters:
        - $ref: "#/components/parameters/child_id"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ReasonRequest"
      responses:
        "200":
          description: Child marked inactive
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Child"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/ForbiddenRole"
        "404":
          $ref: "#/components/responses/NotFound"
...